==> notebooks/01.0_balancing_dataset.ipynb <==
import pandas as pd

# Use a forward slash or escape backslashes to avoid accidental escape sequences like '\f'
df = pd.read_csv("../Data/balanced_f1_dataset.csv")   # change to your actual path if different
df.head()


from sklearn.preprocessing import LabelEncoder
import matplotlib.pyplot as plt
import joblib
le=LabelEncoder()
df["label"]=le.fit_transform(df["winner_name"])
joblib.dump(le, "models/label_encoder.pkl")
print(list(le.classes_))

df['winner_name'].value_counts()

# applying balancing by calculating median count
counts = df['label'].value_counts()
target = int(counts.median())
target

#applying balancing by setting all labels as target count
balanced_df = pd.DataFrame()

for label, count in counts.items():
    class_df = df[df['label'] == label]

    if count > target:
        # Undersample large classes
        class_df = class_df.sample(int(target), random_state=42)
    else:
        # Oversample small classes
        class_df = class_df.sample(int(target), replace=True, random_state=42)

    balanced_df = pd.concat([balanced_df, class_df], ignore_index=True)

# Shuffle final dataset
balanced_df = balanced_df.sample(frac=1, random_state=42).reset_index(drop=True)


balanced_df['winner_name'].value_counts()

balanced_df.to_csv("../Data/balanced_f1_dataset.csv", index=False)
